ENH: Add support for defining which words in the unit should be inflected
Change the notation so that if some words in the unit string shouldn't
be affected by case inflection, the string can be entered in a tuple with
a list of indexes specifying the words that should NOT be inflected. If
the indexes haven't been specified, all words in the string will be
inflected.

Added a new method _add_slots() to the format_numerals_fi that will unwrap
the unit and add the proper case attributes, generating a new LiteralSlot
for each word in the unit string.