ENH: Improve the realization of the change unit type in Finnish
The default realization for the change unit type is now as follows:
...rikosten (suhteellinen) määrä kasvoi/laski viidenneksi eniten/
viisi prosenttia/viidellä (muihin rikostyyppeihin/alueisiin verrattuna)

Note that there is now an inherent assumption that there will be a value
slot before the unit slot. Actually, the two slots could be combined,
but that should probably be done when/if there is some extra time.

The change realizer will add a bunch of text _before_ the value slot,
replace the unit slot, and add a comparison (in case the unit
is a ranking) to the end.