ENH: Disgard facts that are effective repetition
Effective repetition is defined in a bit of a tricky way. In practice,
a fact is repetitive if the only differences are between the *type* of
rank (i.e. one is rank and other is rank_reverse or neither is rank),
or if the difference is only in whether the fact is normalized or not.

Each satellite candidate is compared in this way to all previously
selected satellites and incuded only if neither of the above two cases
apply.

This is likely to be somehow incorrect, but should at least get us
started.

As a side not, this seems like a good place to add support for per-user
acknowledgement for whether the user already likely knows something or
not.