ENH: Remove unnecessary fields from Message
It has become apparent that the Message (rather, Fact) fields what_x, what_type_x,
where_x, where_type_x and when_type_x can be combined to single fields without
the numbered suffix. In other words, we only retain the numbered variants for
when (i.e. when_1 and when_2).

This major changeset removes references to the removed fields and replaces them
with references to the new fields.

Known issue: certain columns (units) no longer get realized correctly. This
needs to be fixed separately. This changeset was pushed to master despite this
known issue as it is so major that it affects almost every facet of the code base.