ENH: Smarter paragraph themes
If a paragraph start to talk about a change from Y1 to Y2,
we also want to allow facts that are about Y1 and Y2 alone.

If the paragraph starts to talk about Y1, we want to allow
facts that are about a change either starting from or ending
in Y1.