ENH: Change NER to use the numeral formatters, add support for {time}
Core entity name resolver decides whether a full or shorter form of the
entity name should be used, after which the actual realization is done by
the language-specific format_numerals_xx.

_time_change_month() and _time_change_year() methods were added to the
Finnish formatter. These are used, when a {time} slot is encountered in
connection to a fact describing change.

Also, realizing the place slots was moved to the numeral formatter
(which should be probably called the slot realizer or something)