ENH: More uniform handling of time interestingness
For years, the interestingness is 1/diff^2 where diff is the difference
from the year of the fact to 2019 (i.e. the next year). This means that
the interestingness for stats from 2018 is 1, for 2017 it's 0.25, for
2016 it's 0.11... etc.

For month M of year Y, the interestingness is allocated linearly between
years Y and Y-1 with the caveat that we consider the year to consist of
13 months so that 2018M12 is the same distance from 2019 as 2018M12 is from
2018M11.