FIX: Incorrect municipality names causing a map problem
The municipality names with letters "ä" had dotless versions in
the REFERENCES variable. This caused the map generator to not realize
that the closest reference is the same as the location for the map
and for these 6 municipalities the resulting map was zoomed in to the
center of the municipality instead of showing the country wide map.