FIX: A minor bug in word order caused by an earlier fix
In the second hunk, the slot needs to be added to idx-1 to end up in the
correct position.
Also, the yearly options in these cases are not needed since
_crime_place_year grouping is only valid for monthly statistics. Changed
the code to raise an AttributeError in the unlikely case that when_type is
something else than "month".

Changed the other Exceptions in the code into AttributeErrors as well,
since it seems to be more descriptive.