FIX: Use Haverstein distance for map distance calcs
Turns out Finland is North, which was a surprise. That in turn meant that the
distances as calculated by euclidean distance on the coordinates are significantly
wrong, especially as the NS-axis distance between two coords is significantly longer
than the EW-distance between two coords in km, for similar distance in coordinates.