FIX: Stop generating unnecessary columns
fetch_crime_data was generating both _rank and _reverse_rank columns for
change data in addition to the _increase_rank and _decrease_rank columns.
The first two shouldn't be used with change data, so they shouldn't be
generated either.

Also set the _rank values for point data to nan if the actual value is
zero. While technically you can have the most crimes even when the value
is zero (tied first place with everybody else), it's not really something
that should be reported.