FIX: Helsinki not considered in landmark generation
This was cause by accidentally dropping the most populous entry from the
municipality list twice, instead of just once. The original first entry
is the country-wide population, which should be dropped, but the next one,
Helsinki was accidentally dropped as well.

Also fixed a minor problem in the output of the prefetch phase.