CLEANUP: Formatting, types and core simplification
This is a large changeset that shouldn't really do any functional
changes. It does five things:

1) Sets up autoformatting/linting with reasonable defaults.
2) Formats the project using those standards. This includes manual
   reformatting where PEP8 errors remained after running isort and
   black.
3) Adds a pre-commit-hook config for automatically running black and
   flake8.
4) Adds more type hints to the code. Type coverage is improved but
   still not 100%. Improving typing is an ongoing process.
5) The eunlg.core package is simplified, with multiple domain modules
   being folded in to a single module called domain.py. This both
   simplifies the code a bit, but also gets rid of some circular
   dependencies that were annoying during the introduction of type
   hints. This code could also use a simplification otherwise, but that
   is going to stay on the TODO list for the immediate future.

Code formatting is done using three libraries:
 - isort is set up to automatically handle the sorting of import
   statements. Run manually as "isort -rc ." in project root.
 - black is set up to autoformat the code using a reasonable default
   values. Run manually as "black ." in project root.
 - flake8 has been set up to automatically lint the project, with a few
   exceptions to e.g. ignoring line lenghts in dictionary files.

All three have been configured so that they should work together, with
the exception of a minor disagreement between isort and black. Run isort
before black.

Project maximum line length is now 120 characters. Both 80 characters
(PEP8 default) and 88 characters (black default) seem unreasonably
strict given the screen sizes and editors of today.

All further commits are expected to run the following before finalizing
the commit:
 $ isort -rc . && black . && flake8 .

Any warning or errors from flake8 are to be viewed as commit-blockers
and need to be fixed before pushing the commit to the origin.

It is strongly recommended that you run "pre-commit install" to set up
the provided pre-commit hooks for running flake8 and black for every
commit.